[DE] Removed cyclic dependency by extracting NetCdfDialog.
