+++ Applications/DataExplorer/DataView/StratView/CMakeLists.txt
@@ -15,7 +15,6 @@ set(HEADERS
 set(SOURCE_DIR_REL ${CMAKE_CURRENT_SOURCE_DIR}/../../../..)
 set(GUI_SOURCE_DIR_REL ${CMAKE_CURRENT_SOURCE_DIR}/../..)
 include_directories(
-    ${CMAKE_CURRENT_BINARY_DIR}
     ${CMAKE_CURRENT_SOURCE_DIR}
     ${SOURCE_DIR_REL}/BaseLib
     ${SOURCE_DIR_REL}/GeoLib
