+++ Applications/DataExplorer/NetCdfDialog/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_library(NetCdfDialogLib NetCdfConfigureDialog.cpp NetCdfConfigureDialog.h)
+target_link_libraries(NetCdfDialogLib PUBLIC Qt5::Widgets PRIVATE MathLib)
+set_property(TARGET NetCdfDialogLib PROPERTY FOLDER "DataExplorer")
